LPDEV-57 Merge

+++ src/Lykke.Pay.Invoice/Models/GridModel.cs
@@ -13,7 +13,7 @@ namespace Lykke.Pay.Invoice.Models
         public GridHeader Header { get; set; }
         public GridFilter Filter { get; set; }
         public int Page { get; set; }
-        public int PageCount { get; set; }
+        public decimal PageCount { get; set; }
         
         public IList<InvoiceModel> Data { get; set; }
     }
